create

+++ src/electrum/electrs.h
@@ -1,6 +1,3 @@
-// Copyright (c) 2019 The Bitcoin Unlimited developers
-// Distributed under the MIT software license, see the accompanying
-// file COPYING or http://www.opensource.org/licenses/mit-license.php.
 #ifndef ELECTRUM_ELECTRS_H
 #define ELECTRUM_ELECTRS_H
 
@@ -16,7 +13,7 @@ namespace electrum {
 std::string electrs_path();
 std::vector<std::string> electrs_args(int rpcport, const std::string& network);
 
-std::map<std::string, int64_t> fetch_electrs_info();
+std::map<std::string, int> fetch_electrs_info();
 
 } // ns electrum
